refactor: remove FloatType from examples & doc

+++ honeycomb-core/src/utils/generation.rs
@@ -512,6 +512,7 @@ impl<T: CoordsFloat> GridBuilder<T> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::FloatType;
 
     #[test]
     fn build_nc_lpc_l() {
@@ -603,7 +604,7 @@ mod tests {
 
     #[test]
     fn square_cmap2_correctness() {
-        let cmap: CMap2<f64> = GridBuilder::unit_squares(2).build2().unwrap();
+        let cmap: CMap2<FloatType> = GridBuilder::unit_squares(2).build2().unwrap();
 
         // hardcoded because using a generic loop & dim would just mean
         // reusing the same pattern as the one used during construction
@@ -704,7 +705,7 @@ mod tests {
     #[allow(clippy::too_many_lines)]
     #[test]
     fn splitsquare_cmap2_correctness() {
-        let cmap: CMap2<f64> = GridBuilder::split_unit_squares(2).build2().unwrap();
+        let cmap: CMap2<FloatType> = GridBuilder::split_unit_squares(2).build2().unwrap();
 
         // hardcoded because using a generic loop & dim would just mean
         // reusing the same pattern as the one used during construction
